implemented bind maps to improve performance of binding

+++ src/binder/bind-null.cc
@@ -2,8 +2,8 @@
 // If no index is given, the next anonymous index is used.
 // If an error occurs, error is set to an appropriately descriptive string.
 
-void Binder::BindNull(unsigned int index) {
+void Binder::BindNull(int index) {
 	if (!index) {index = NextAnonIndex();}
-	int status = sqlite3_bind_null(handle, (int)index);
+	int status = sqlite3_bind_null(handle, index);
 	SetBindingError(status);
 }
