+++ src/binder/bind-string.cc
@@ -2,11 +2,11 @@
 // If no index is given, the next anonymous index is used.
 // If an error occurs, error is set to an appropriately descriptive string.
 
-void Binder::BindString(v8::Local<v8::String> value, unsigned int index) {
+void Binder::BindString(v8::Local<v8::String> value, int index) {
 	if (!index) {index = NextAnonIndex();}
 	
 	v8::String::Value utf16(value);
-	int status = sqlite3_bind_text16(handle, (int)index, *utf16, utf16.length() * sizeof (uint16_t), SQLITE_TRANSIENT);
+	int status = sqlite3_bind_text16(handle, index, *utf16, utf16.length() * sizeof (uint16_t), SQLITE_TRANSIENT);
 	
 	SetBindingError(status);
 }
